code/pyproject.toml
[project]
name = "ege-rag-workspace"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "numpy>=1.26",
    "pandas>=2.2.3",
    "qdrant-client>=1.12.0",
    "fastembed>=0.3.6",
    "groq>=0.11.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "langgraph>=1.1.8",
    "langchain-core>=1.3.0",
    "langchain-mcp-adapters>=0.2.2",
    "fastmcp>=3.2.4",
    "pydantic-settings>=2.2",
    # Platform integration (Phase 1+)
    "asyncpg>=0.30",
    "alembic>=1.14",
    "sqlalchemy>=2.0",
    "httpx>=0.27",
    "tenacity>=9.0",
    "aiokafka>=0.12",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.6.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.15.12",
    "respx>=0.21",
    "testcontainers[postgresql,kafka]>=4.9",
]
eval = [
    "deepeval>=2.0.0",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
ignore = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pandas-stubs>=2.2.0",
]
